+++ cypress/integration/display-plate.spec.ts
@@ -1,21 +1,15 @@
 import {
   BARCODE_1,
   BARCODE_2,
-  BLUE_COLOR_1,
   checkAllWells,
-  checkColumn,
-  DARK_COLOR,
-  DARK_GREEN_COLOR,
-  DARK_YELLOW_COLOR,
+  checkPlate,
+  clickOnBootstrapCheckbox,
   mustBeReadonlyPlate,
-  PINK_RED_COLOR,
-  PURPLE_COLOR,
-  RED_COLOR,
   selectAllWells,
   selectObject,
+  selectPlate,
   TURQUOISE_COLOR,
   WHITE_COLOR,
-  YELLOW_COLOR,
 } from '../support/plate-models';
 
 describe('Display plate', () => {
@@ -57,15 +51,17 @@ describe('Display plate', () => {
   it('should override the default truncate logic', () => {
     selectPlate(BARCODE_1);
     checkPlate(BARCODE_1, [0, 11]);
-    cy.get('#override-truncate').click().should('be.checked');
+    clickOnBootstrapCheckbox(cy.get('#override-truncate'));
+    cy.get('#override-truncate').should('be.checked');
     checkAllWells('Over-' + BARCODE_1.substring(0, 7));
   });
 
   it('should take control of the change detection in plate-maker', () => {
-    cy.get('#control-tick').click().should('be.checked');
+    clickOnBootstrapCheckbox(cy.get('#control-tick'));
+    cy.get('#control-tick').should('be.checked');
     selectPlate(BARCODE_2);
     checkAllWells(BARCODE_2.substring(0, 10) + '..');
-    cy.contains('button', 'Tick').click();
+    cy.contains('button', 'tick').click();
     checkPlate(BARCODE_2, [2]);
     selectPlate(BARCODE_1);
     checkAllWells(BARCODE_1.substring(0, 10) + '..');
@@ -73,9 +69,10 @@ describe('Display plate', () => {
     checkPlate(BARCODE_2, [2]);
     selectPlate(BARCODE_1);
     checkAllWells(BARCODE_1.substring(0, 10) + '..');
-    cy.contains('button', 'Tick').click();
+    cy.contains('button', 'tick').click();
     checkPlate(BARCODE_1, [4]);
-    cy.get('#control-tick').click().should('not.be.checked');
+    clickOnBootstrapCheckbox(cy.get('#control-tick'));
+    cy.get('#control-tick').should('not.be.checked');
     selectPlate(BARCODE_2);
     checkPlate(BARCODE_2, [2]);
     selectPlate(BARCODE_1);
@@ -84,25 +81,3 @@ describe('Display plate', () => {
 
 });
 
-
-function selectPlate(barcode: string) {
-  return cy.get('.barcode').contains(barcode).click();
-}
-
-function checkPlate(barcode: string, columns= [0, 6, 11]) {
-  const color = {
-    0: TURQUOISE_COLOR,
-    1: BLUE_COLOR_1,
-    2: PURPLE_COLOR,
-    4: RED_COLOR,
-    5: DARK_YELLOW_COLOR,
-    6: DARK_COLOR,
-    9: PINK_RED_COLOR,
-    10: DARK_GREEN_COLOR,
-    11: YELLOW_COLOR,
-  };
-  for (const colIdx of columns) {
-    const content = barcode + colIdx;
-    checkColumn(colIdx + 1, '*' + (content).substring(content.length - 11, content.length), [8, 12], color[colIdx]);
-  }
-}
