+++ cypress/support/plate-models.ts
@@ -1,3 +1,5 @@
+import Chainable = Cypress.Chainable;
+
 export const BARCODE_1 = 'BjDn8z9499rzIxEozO9m';
 export const BARCODE_2 = 'cyVm8CccTNUi4EuKkwbM';
 
@@ -40,6 +42,13 @@ export class Dimension {
   }
 }
 
+export function clickOnBootstrapCheckbox(chainable: Chainable) {
+  // because Bootstrap use before and after pseudo-element to adjust label size that has initially 0*0 px
+  // here we force the click because the element 'input' is being covered by another element 'label'
+  // https://docs.cypress.io/guides/references/error-messages.html#cy-failed-because-the-element-cannot-be-interacted-with
+  chainable.click({force: true});
+}
+
 export function selectWell(row: number, col: number, dimension = new Dimension(96)) {
   cy.get('dts-select-container').then((containers) => {
     const wellWidth = (containers[0].offsetWidth - 25) / dimension.cols();
@@ -65,19 +74,41 @@ export function checkWell(row: number, col: number, content: string | string[], 
   cy.get('.plate-col').eq((row - 1) * nbCols + (col - 1)).then(well => checkWellContent(well, content, bgColor));
 }
 
-export function checkColumn(col: number, content: string, shape: number[], bgColor?: string) {
-  for (let i = col - 1; i < shape[0] * shape[1]; i += 2 * shape[1]) {  // check every other well to speed up tests a bit
-    cy.get('.plate-col').eq(i).then(well => checkWellContent(well, content, bgColor));
-    cy.wait(0);
-  }
-}
-
 export function checkAllWells(content: string | string[], bgColor?: string) {
   cy.get('.plate-col').then((well: JQuery<HTMLElement>) => checkWellContent(well, content, bgColor));
 }
 
 export function mustBeReadonlyPlate() {
   cy.get('.plate-col').then((well) => expect(well).to.css('cursor', 'not-allowed'));
+}
+
+export function selectPlate(barcode: string) {
+  return cy.get('.barcode').contains(barcode).click();
+}
+
+export function checkPlate(barcode: string, columns = [0, 6, 11]) {
+  const color = {
+    0: TURQUOISE_COLOR,
+    1: BLUE_COLOR_1,
+    2: PURPLE_COLOR,
+    4: RED_COLOR,
+    5: DARK_YELLOW_COLOR,
+    6: DARK_COLOR,
+    9: PINK_RED_COLOR,
+    10: DARK_GREEN_COLOR,
+    11: YELLOW_COLOR,
+  };
+  for (const colIdx of columns) {
+    const content = barcode + colIdx;
+    checkColumn(colIdx + 1, '*' + (content).substring(content.length - 11, content.length), [8, 12], color[colIdx]);
+  }
+}
+
+function checkColumn(col: number, content: string, shape: number[], bgColor?: string) {
+  for (let i = col - 1; i < shape[0] * shape[1]; i += 2 * shape[1]) {  // check every other well to speed up tests a bit
+    cy.get('.plate-col').eq(i).then(well => checkWellContent(well, content, bgColor));
+    cy.wait(0);
+  }
 }
 
 function checkWellContent(well: JQuery<HTMLElement>, content: string | string[], bgColor?: string) {
