+++ projects/@abolis/plate-maker/ng-package.json
@@ -6,7 +6,8 @@
     "umdModuleIds": {
       "ngx-drag-to-select": "ngx-drag-to-select",
       "ngx-popper": "ngx-popper",
-      "angular-animations": "angular-animations"
+      "angular-animations": "angular-animations",
+      "string-algorithms": "string-algorithms"
     }
   }
 }
