+++ projects/app/src/app/create-plate-app/create-plate-app.component.html
@@ -38,7 +38,7 @@
       <div id="printed" class="text-center animated fadeIn p-3" *ngIf="savedWells?.length">
         <div>Wells that could have been saved to a backend in real life</div>
         <div *ngFor="let well of savedWells" class="saved-well">
-          ({{well.listDisplay}}
+          {{well.listDisplay}}
         </div>
       </div>
     </div>
