+++ projects/app/src/app/display-plate-app/display-plate-app.component.html
@@ -15,7 +15,28 @@
         <div class="card-header elegant-color white-text text-center font-weight-bold">Barcode(s)</div>
         <div class="card-body">
           <div class="d-flex justify-content-between pb-2">
-            <div class="mt-1 p-0 font-weight-bold"></div>
+            <div class="p-0">
+              <div class="custom-control custom-checkbox">
+                <input type="checkbox"
+                       class="custom-control-input"
+                       id="override-truncate"
+                       [formControl]="truncateControl">
+                <label class="custom-control-label" for="override-truncate">Override truncate</label>
+              </div>
+              <div class="d-flex flex-row mt-1">
+                <div class="custom-control custom-checkbox">
+                  <input type="checkbox" class="custom-control-input" id="control-tick" [formControl]="tickControl">
+                  <label class="custom-control-label" for="control-tick">Control tick</label>
+                </div>
+                <div>
+                  <button class="btn btn-sm btn-dark px-3 py-1 my-0"
+                          [disabled]="!tickControl.value"
+                          (click)="tickEmitter.emit()">
+                    tick
+                  </button>
+                </div>
+              </div>
+            </div>
             <div class="p-0">
               <div class="md-form form-sm m-0 d-inline-block">
                 <input type="text"
@@ -27,19 +48,6 @@
               <span class="badge badge-light z-depth-0 rounded px-0 pl-2 ml-2">
                 Count <span class="elegant-color white-text rounded ml-1 p-1">{{displayedBarcodes.length}}</span>
               </span>
-              <div>
-                <label for="override-truncate">Override truncate</label>
-                <input type="checkbox" id="override-truncate" [formControl]="truncateControl">
-              </div>
-              <div>
-                <label for="control-tick">Control tick</label>
-                <input type="checkbox" id="control-tick" [formControl]="tickControl">
-                <button [disabled]="!tickControl.value"
-                        class="btn btn-sm btn-dark"
-                        (click)="tickEmitter.emit()">
-                  Tick
-                </button>
-              </div>
             </div>
           </div>
           <div class="scrollable p-2" *ngIf="displayedBarcodes.length">
